[wip] Make `TWPrivateKey` and `TWPublicKey` available for end-user
* Move `TWPrivateKey` and `TWPublicKey` to the `tw_keypair::tw` module
* Refactor `TWPublicKey` by making it enum
* TODO add `TWKeyPair`
