+++ rust/tw_keypair/src/lib.rs
@@ -4,9 +4,8 @@
 // terms governing use, modification, and redistribution, is contained in the
 // file LICENSE at the root of the source code distribution tree.
 
-extern crate core;
-
 pub mod ffi;
+pub mod tw;
 pub mod secp256k1;
 pub mod traits;
