+++ rust/tw_keypair/src/tw/mod.rs
@@ -1,0 +1,11 @@
+// Copyright © 2017-2023 Trust Wallet.
+//
+// This file is part of Trust. The full Trust copyright notice, including
+// terms governing use, modification, and redistribution, is contained in the
+// file LICENSE at the root of the source code distribution tree.
+
+mod private;
+mod public;
+
+pub use private::TWPrivateKey;
+pub use public::TWPublicKey;
